requirements.txt
Django==2.0.5
djangorestframework==3.8.2
Faker==0.8.15
python-dateutil==2.7.3
pytz==2018.4
six==1.11.0
text-unidecode==1.2
wheel==0.24.0
